more progress through the guide

+++ GuidedTour.playground/section-11.swift
@@ -3,8 +3,8 @@ let oranges = 5
 let appleSummary = "I have \(apples) apples."
 let fruitSummary = "I have \(apples + oranges) pieces of fruit."
 
-let height:Double = 68
+let height:Float = 68
 let name = "Andrew"
 let myHeight = "\(name)'s height is \(height + 0.5) inches"
-
+//this is an example of string interpolation and doing math in string interpolation.
